Fix SQLITE_OMIT_WSD feature regression
This can cause multiple threads to initialize the builtin functions at the same time when using SQLITE_OMIT_WSD feature along with SQLITE_THREADSAFE = 0 (no mutex) so that each sqlite connection acts like it is in a separate process for read scalability reasons. For this case to work, there must be no shared globals between threads.